api/requirements.txt
fastapi
uvicorn[standard]
pyserial
opencv-python-headless
pydantic
websockets
